refactor header links to own component
