+++ src/components/header/AppHeaderBottomBar.vue
@@ -1,129 +1,20 @@
 <script setup lang="ts">
-import { computed, ref, watch } from 'vue';
-import { useRoute, useRouter } from 'vue-router';
+import { ref } from 'vue';
 import { useQuasar } from 'quasar';
 
-import LanguageSwitcherModal from 'components/header/LanguageSwitcherModal.vue';
 import AppHeaderButton from 'components/header/AppHeaderButton.vue';
 import AppHeaderWallet from 'components/header/AppHeaderWallet.vue';
+import AppHeaderLinks from 'components/header/AppHeaderLinks.vue';
 
-const $router = useRouter();
-const $route = useRoute();
 const $q = useQuasar();
-
-// eztodo get these from somewhere else?
-const TELOSCAN_MAINNET_URL = 'https://teloscan.io';
-const TELOSCAN_TESTNET_URL = 'https://testnet.teloscan.io';
-
-// eztodo i18n
-const blockchainSubmenuItems = [
-    { name: 'transactions', label: 'Transactions' },
-    { name: 'blocks', label: 'Blocks' },
-];
-
-// eztodo handle testnet, use chain settings for links, i18n
-const developersSubmenuItems = [
-    { url: 'https://api.teloscan.io/swagger/', label: 'API Documentation' },
-    { url: 'https://sourcify.dev/', label: 'Verify Contract (Sourcify)' },
-];
-
-// eztodo use chain settings for link, i18n
-const walletMenuItem = {
-    url: 'https://wallet.telos.net/',
-    label: 'Telos Wallet',
-};
-
-// eztodo use chain settings, i18n
-const moreSubmenuItems = {
-    internal: [
-        { name: 'export', label: 'CSV Export' },
-    ],
-    external: [
-        {
-            url: 'https://www.telos.net/ecosystem',
-            label: 'Telos Ecosystem',
-        },
-    ],
-};
-
-// eztodo i18n
-const networksMenuItems = {
-    mainnet: [{
-        url: TELOSCAN_MAINNET_URL,
-        label: 'Telos Mainnet',
-    }],
-    testnet: [{
-        url: TELOSCAN_TESTNET_URL,
-        label: 'Telos Testnet',
-    }],
-};
 
 // data
 const menuBottomBarHidden = ref(false);
-const showLanguageSwitcher = ref(false);
 const menuVisibleMobile = ref(false);
-const blockchainMenuExpandedMobile = ref(false);
-const developersMenuExpandedMobile = ref(false);
-const moreMenuExpandedMobile = ref(false);
-const networkMenuExpandedMobile = ref(false);
-
-// watchers
-watch(() => $q.screen.gt.md, () => {
-    // eztodo fix this
-    closeAllMenus();
-});
-
-// computed
-const highlightBlockchainMenuItem = computed(() => blockchainSubmenuItems.some(({ name }) => name === $route.name));
-const highlightMoreMenuItem = computed(() => moreSubmenuItems.internal.some(({ name }) => name === $route.name));
-
 
 // methods
 function scrollHandler(info: { direction: string; }) {
     menuBottomBarHidden.value = info.direction === 'down';
-}
-
-function blurActiveElement() {
-    (document.activeElement as HTMLElement | null)?.blur();
-}
-
-function closeAllMenus() {
-    menuVisibleMobile.value = false;
-    blockchainMenuExpandedMobile.value = false;
-    developersMenuExpandedMobile.value = false;
-    moreMenuExpandedMobile.value = false;
-    networkMenuExpandedMobile.value = false;
-}
-
-function toggleDarkMode() {
-    $q.dark.toggle();
-    localStorage.setItem('darkModeEnabled', $q.dark.isActive.toString());
-}
-
-function getIsCurrentNetworkMenuItem(url: string) {
-    if (url === TELOSCAN_MAINNET_URL) {
-        return Number(process.env.NETWORK_EVM_CHAIN_ID) === 40;
-    }
-
-    if (url === TELOSCAN_TESTNET_URL) {
-        return Number(process.env.NETWORK_EVM_CHAIN_ID) === 41;
-    }
-
-    return false;
-}
-
-function goTo(to: string | { name: string }) {
-    // mobileMenuIsOpen.value = false;
-    blurActiveElement();
-    closeAllMenus();
-
-    const httpsRegex = /^https/;
-    if (typeof to === 'string' && httpsRegex.test(to)) {
-        window.open(to, '_blank');
-        return;
-    }
-
-    $router.push(to);
 }
 </script>
 
@@ -150,7 +41,7 @@ function goTo(to: string | { name: string }) {
                 Teloscan
             </span>
         </router-link>
-        <!-- eztodo aria and kb and break into components -->
+
         <nav class="c-header-bottom-bar__right-container">
             <AppHeaderWallet v-if="$q.screen.lt.md" class="q-mr-sm" />
 
@@ -163,294 +54,22 @@ function goTo(to: string | { name: string }) {
                 <q-icon name="menu" size="24px" />
             </AppHeaderButton>
 
-            <ul
-                :class="{
-                    'c-header-bottom-bar__menu-ul': true,
-                    'c-header-bottom-bar__menu-ul--visible-mobile shadow-2': menuVisibleMobile && $q.screen.lt.md,
-                }"
-            >
-                <li
-                    :class="{
-                        'c-header-bottom-bar__menu-li': true,
-                        'c-header-bottom-bar__menu-li--current': $route.name === 'home',
-                    }"
-                    tabindex="0"
-                    role="link"
-                    @click="goTo({ name: 'home' })"
-                    @keydown.enter="goTo({ name: 'home' })"
-                >
-                    Home
-                </li>
-
-                <li
-                    :class="{
-                        'c-header-bottom-bar__menu-li c-header-bottom-bar__menu-li--expandable': true,
-                        'c-header-bottom-bar__menu-li--current': highlightBlockchainMenuItem,
-                        'c-header-bottom-bar__menu-li--expanded-mobile': blockchainMenuExpandedMobile,
-                    }"
-                    tabindex="0"
-                    :aria-expanded="$q.screen.lt.md ? blockchainMenuExpandedMobile : undefined"
-                    :aria-controls="$q.screen.lt.md ? 'app-header-blockchain-submenu-ul' : undefined"
-                    :aria-labelledby="$q.screen.lt.md ? 'app-header-blockchain-submenu-label' : undefined"
-                    @mouseleave="blurActiveElement"
-                    @click="blockchainMenuExpandedMobile = !blockchainMenuExpandedMobile"
-                    @keydown.enter="blockchainMenuExpandedMobile = !blockchainMenuExpandedMobile"
-                >
-                    <div id="app-header-blockchain-submenu-label" class="c-header-bottom-bar__menu-li-text">
-                        <!-- eztodo i18n -->
-                        Blockchain
-                        <q-icon name="fas fa-chevron-down" size="12px" />
-                    </div>
-
-                    <ul
-                        id="app-header-blockchain-submenu-ul"
-                        :class="{
-                            'c-header-bottom-bar__submenu-ul': true,
-                            'shadow-2': $q.screen.gt.md,
-                        }"
-                    >
-                        <li
-                            v-for="item in blockchainSubmenuItems"
-                            :key="`blockchain-submenu-item-${item.name}`"
-                            :class="{
-                                'c-header-bottom-bar__submenu-li': true,
-                                'c-header-bottom-bar__submenu-li--current': $route.name === item.name,
-                            }"
-                            tabindex="0"
-                            role="link"
-                            @click="goTo({ name: item.name })"
-                            @keydown.enter="goTo({ name: item.name })"
-                        >
-                            {{ item.label }}
-                        </li>
-                    </ul>
-                </li>
-
-                <li
-                    :class="{
-                        'c-header-bottom-bar__menu-li c-header-bottom-bar__menu-li--expandable': true,
-                        'c-header-bottom-bar__menu-li--expanded-mobile': developersMenuExpandedMobile,
-                    }"
-                    tabindex="0"
-                    :aria-expanded="$q.screen.lt.md ? developersMenuExpandedMobile : undefined"
-                    :aria-controls="$q.screen.lt.md ? 'app-header-developers-submenu-ul' : undefined"
-                    :aria-labelledby="$q.screen.lt.md ? 'app-header-developers-submenu-label' : undefined"
-                    @mouseleave="blurActiveElement"
-                    @click="developersMenuExpandedMobile = !developersMenuExpandedMobile"
-                    @keydown.enter="developersMenuExpandedMobile = !developersMenuExpandedMobile"
-                >
-                    <div id="app-header-developers-submenu-label" class="c-header-bottom-bar__menu-li-text">
-                        <!-- eztodo i18n -->
-                        Developers
-                        <q-icon name="fas fa-chevron-down" size="12px" />
-                    </div>
-
-                    <ul
-                        id="app-header-developers-submenu-ul"
-                        :class="{
-                            'c-header-bottom-bar__submenu-ul': true,
-                            'shadow-2': $q.screen.gt.md,
-                        }"
-                    >
-                        <li
-                            v-for="item in developersSubmenuItems"
-                            :key="`developer-submenu-item-${item.label}`"
-                            class="c-header-bottom-bar__submenu-li"
-                            tabindex="0"
-                            role="link"
-                            @click="goTo(item.url)"
-                            @keydown.enter="goTo(item.url)"
-                        >
-                            <div class="u-flex--center-y">
-                                {{ item.label }}
-                                <q-icon name="fas fa-external-link-alt" size="12px" class="q-ml-sm" />
-                            </div>
-                        </li>
-                    </ul>
-                </li>
-
-                <li
-                    class="c-header-bottom-bar__menu-li"
-                    tabindex="0"
-                    role="link"
-                    @click="goTo(walletMenuItem.url)"
-                    @keydown.enter="goTo(walletMenuItem.url)"
-                >
-                    {{ walletMenuItem.label }}
-                </li>
-
-                <li
-                    :class="{
-                        'c-header-bottom-bar__menu-li c-header-bottom-bar__menu-li--expandable': true,
-                        'c-header-bottom-bar__menu-li--current': highlightMoreMenuItem,
-                        'c-header-bottom-bar__menu-li--expanded-mobile': moreMenuExpandedMobile,
-                    }"
-                    tabindex="0"
-                    :aria-expanded="$q.screen.lt.md ? moreMenuExpandedMobile : undefined"
-                    :aria-controls="$q.screen.lt.md ? 'app-header-more-submenu-ul' : undefined"
-                    :aria-labelledby="$q.screen.lt.md ? 'app-header-more-submenu-label' : undefined"
-                    @mouseleave="blurActiveElement"
-                    @click="moreMenuExpandedMobile = !moreMenuExpandedMobile"
-                    @keydown.enter="moreMenuExpandedMobile = !moreMenuExpandedMobile"
-                >
-                    <div id="app-header-more-submenu-label" class="c-header-bottom-bar__menu-li-text">
-                        <!-- eztodo i18n -->
-                        More
-                        <q-icon name="fas fa-chevron-down" size="12px" />
-                    </div>
-
-                    <ul
-                        id="app-header-more-submenu-ul"
-                        :class="{
-                            'c-header-bottom-bar__submenu-ul c-header-bottom-bar__submenu-ul--rightmost': true,
-                            'shadow-2': $q.screen.gt.md,
-                        }"
-                    >
-                        <li
-                            class="c-header-bottom-bar__submenu-li"
-                            tabindex="0"
-                            role="button"
-                            :aria-label="'eztodo'"
-                            @keydown.enter="showLanguageSwitcher = true"
-                            @click="showLanguageSwitcher = true"
-                        >
-                            <div class="u-flex--center-y">
-                                {{ $t('global.language') }}
-                                <q-icon name="fas fa-language" size="16px" class="q-ml-sm" />
-                            </div>
-                        </li>
-
-                        <li
-                            v-for="item in moreSubmenuItems.internal"
-                            :key="`more-submenu-item-internal-${item.name}`"
-                            :class="{
-                                'c-header-bottom-bar__submenu-li': true,
-                                'c-header-bottom-bar__submenu-li--current': $route.name === item.name,
-                            }"
-                            tabindex="0"
-                            role="link"
-                            @click="goTo({ name: item.name })"
-                            @keydown.enter="goTo({ name: item.name })"
-                        >
-                            {{ item.label }}
-                        </li>
-
-                        <li
-                            v-for="item in moreSubmenuItems.external"
-                            :key="`more-submenu-item-external-${item.label}`"
-                            class="c-header-bottom-bar__submenu-li"
-                            tabindex="0"
-                            role="link"
-                            @click="goTo(item.url)"
-                            @keydown.enter="goTo(item.url)"
-                        >
-                            <div class="u-flex--center-y">
-                                {{ item.label }}
-                                <q-icon name="fas fa-external-link-alt" size="12px" class="q-ml-sm" />
-                            </div>
-                        </li>
-                    </ul>
-                </li>
-
-                <li
-                    v-if="$q.screen.lt.md"
-                    :class="{
-                        'c-header-bottom-bar__menu-li c-header-bottom-bar__menu-li--expandable': true,
-                        'c-header-bottom-bar__menu-li--expanded-mobile': networkMenuExpandedMobile,
-                    }"
-                    tabindex="0"
-                    :aria-expanded="$q.screen.lt.md ? networkMenuExpandedMobile : undefined"
-                    :aria-controls="$q.screen.lt.md ? 'app-header-network-submenu-ul' : undefined"
-                    :aria-labelledby="$q.screen.lt.md ? 'app-header-network-submenu-label' : undefined"
-                    @mouseleave="blurActiveElement"
-                    @click="networkMenuExpandedMobile = !networkMenuExpandedMobile"
-                    @keydown.enter="networkMenuExpandedMobile = !networkMenuExpandedMobile"
-                >
-                    <div id="app-header-network-submenu-label" class="c-header-bottom-bar__menu-li-text">
-                        <!-- eztodo i18n -->
-                        Network
-                        <q-icon name="fas fa-chevron-down" size="12px" />
-                    </div>
-
-                    <ul
-                        id="app-header-network-submenu-ul"
-                        :class="{
-                            'c-header-bottom-bar__submenu-ul': true,
-                            'shadow-2': $q.screen.gt.md,
-                        }"
-                    >
-                        <li
-                            v-for="item in networksMenuItems.mainnet"
-                            :key="`networks-submenu-item-mainnet-${item.label}`"
-                            :class="{
-                                'c-header-bottom-bar__submenu-li': true,
-                                'c-header-bottom-bar__submenu-li--current': getIsCurrentNetworkMenuItem(item.url),
-                            }"
-                            tabindex="0"
-                            role="link"
-                            @click="goTo(item.url)"
-                            @keydown.enter="goTo(item.url)"
-                        >
-                            {{ item.label }}
-                        </li>
-
-                        <q-separator />
-
-                        <li
-                            v-for="item in networksMenuItems.testnet"
-                            :key="`networks-submenu-item-testnet-${item.label}`"
-                            :class="{
-                                'c-header-bottom-bar__submenu-li': true,
-                                'c-header-bottom-bar__submenu-li--current': getIsCurrentNetworkMenuItem(item.url),
-                            }"
-                            tabindex="0"
-                            role="link"
-                            @click="goTo(item.url)"
-                            @keydown.enter="goTo(item.url)"
-                        >
-                            {{ item.label }}
-                        </li>
-                    </ul>
-                </li>
-
-                <li v-if="$q.screen.lt.md" class="c-header-bottom-bar__menu-li">
-                    <AppHeaderButton
-                        text-color="default"
-                        class="c-header-bottom-bar__theme-toggle"
-                        @click="toggleDarkMode"
-                    >
-                        <!-- eztodo i18n -->
-                        <q-icon
-                            :name="`far fa-${$q.dark.isActive ? 'moon' : 'sun'}`"
-                            size="14px"
-                            color="primary"
-                            class="q-mr-sm"
-                        />
-                        Switch to light theme
-                    </AppHeaderButton>
-                </li>
-            </ul>
+            <AppHeaderLinks
+                :menu-visible-mobile="menuVisibleMobile"
+                @close-menu="menuVisibleMobile = false"
+            />
         </nav>
     </div>
 </div>
 
 <q-scroll-observer @scroll="scrollHandler" />
-
-<LanguageSwitcherModal :show="showLanguageSwitcher" @hide="showLanguageSwitcher = false" />
 </template>
 
 <style lang="scss">
 .c-header-bottom-bar {
     $this: &;
+
     padding: 0 12px;
-
-    // other CSS vars defined in AppHeader.vue
-    --highlight-color: #{$grey-3};
-
-    @at-root .body--dark & {
-        --highlight-color: #{$grey-9};
-    }
-
     position: absolute;
     top: var(--top-bar-height);
     left: 0;
@@ -512,152 +131,6 @@ function goTo(to: string | { name: string }) {
         @media screen and (min-width: $breakpoint-md-min) {
             display: none;
         }
-    }
-
-    &__menu-ul {
-        display: none;
-        height: 100%;
-        list-style: none;
-        flex-direction: row;
-        margin: 0;
-        padding: 0;
-
-        &--visible-mobile {
-            position: absolute;
-            display: flex;
-            left: 0;
-            right: 0;
-            flex-direction: column;
-            top: 52px;
-            height: max-content;
-            background-color: var(--background-color);
-
-            #{$this}__menu-li--expandable {
-                justify-content: space-between;
-            }
-
-            #{$this}__menu-li--expanded-mobile {
-                #{$this}__submenu-ul {
-                    display: block;
-                    width: 100%;
-                    border: 1px solid var(--border-color);
-                    border-radius: 8px;
-                    padding: 12px;
-                }
-            }
-        }
-
-        @media screen and (min-width: $breakpoint-md-min) {
-            display: flex;
-        }
-    }
-
-    &__menu-li {
-        padding: 8px 16px;
-        display: flex;
-        align-items: center;
-        user-select: none;
-        cursor: pointer;
-        font-weight: 500;
-
-        &:not(&--expandable):hover {
-            color: $primary;
-        }
-
-        @media screen and (min-width: $breakpoint-md-min) {
-            padding: 8px;
-        }
-
-        &--expandable {
-            flex-direction: column;
-            align-items: flex-start;
-            gap: 8px;
-
-            @media screen and (min-width: $breakpoint-md-min) {
-                position: relative;
-                cursor: default;
-                gap: 4px;
-                align-items: center;
-                flex-direction: row;
-
-                &:hover,
-                &:focus,
-                &:focus-within {
-                    .c-header-bottom-bar__submenu-ul {
-                        opacity: 1;
-                        visibility: visible;
-                        transform: translateY(0);
-                    }
-                }
-            }
-        }
-
-        &--current {
-            color: $primary;
-        }
-    }
-
-    &__menu-li-text {
-        display: flex;
-        align-items: center;
-        justify-content: space-between;
-        width: 100%;
-
-        @media screen and (min-width: $breakpoint-md-min) {
-            gap: 8px;
-        }
-    }
-
-    &__submenu-ul {
-        transition: 0.2s ease all;
-
-        display: none;
-        z-index: 999;
-        list-style: none;
-        color: var(--text-color);
-        font-weight: normal;
-
-        @media screen and (min-width: $breakpoint-md-min) {
-            display: block;
-            opacity: 0;
-            padding: 12px;
-            visibility: hidden;
-            transform: translateY(8px);
-            position: absolute;
-            top: calc(100% - 1px);
-            left: 0;
-            justify-content: space-between;
-            border-radius: 0 0 4px 4px;
-            background-color: var(--background-color);
-            border-top: 2px solid $primary;
-
-            &--rightmost {
-                right: 0;
-                left: unset;
-            }
-        }
-    }
-
-    &__submenu-li {
-        padding: 8px;
-        cursor: pointer;
-        user-select: none;
-        transition: 0.2s ease all;
-        border-radius: 8px;
-        min-width: max-content;
-
-        &:hover,
-        &:active {
-            background-color: var(--highlight-color);
-        }
-
-        &--current {
-            color: $primary;
-        }
-    }
-
-    &__theme-toggle {
-        width: 100%;
     }
 }
 </style>
